refactor: extract FullscreenChartFrame from the two fullscreen charts
FullscreenLineChart and FullscreenDualAxisLineChart were ~90% identical
chrome: the fullscreen icon button, the landscape orientation lock +
hidden system bars DisposableEffect, the Dialog, the BoxWithConstraints
height math, and the back button. Moved all of it into a reusable
FullscreenChartFrame that takes an inline slot and a fullscreen slot
(receiving the computed chart height). Both wrappers are now thin call
sites. No behavioural change.
