+++ app/src/main/java/com/matedroid/ui/components/FullscreenLineChart.kt
@@ -1,46 +1,10 @@
 package com.matedroid.ui.components
 
-import android.app.Activity
-import android.content.pm.ActivityInfo
-import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Fullscreen
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalView
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import com.matedroid.R
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import androidx.compose.ui.window.SecureFlagPolicy
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.WindowInsetsControllerCompat
 
 /**
  * A line chart component with fullscreen capability.
@@ -65,134 +29,26 @@ fun FullscreenLineChart(
 ) {
     if (data.size < 2) return
 
-    var isFullscreen by rememberSaveable { mutableStateOf(false) }
-    val context = LocalContext.current
-    val activity = context as? Activity
-
-    Box(modifier = modifier) {
-        // Regular chart
-        OptimizedLineChart(
-            data = data,
-            color = color,
-            unit = unit,
-            showZeroLine = showZeroLine,
-            fixedMinMax = fixedMinMax,
-            timeLabels = timeLabels,
-            convertValue = convertValue,
-            externalSelectedFraction = externalSelectedFraction,
-            onXSelected = onXSelected,
-            fractionToTimeLabel = fractionToTimeLabel,
-            annotationRanges = annotationRanges,
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        // Fullscreen icon button in lower-right corner
-        Box(
-            modifier = Modifier
-                .align(Alignment.BottomEnd)
-                .padding(4.dp)
-                .size(28.dp)
-                .clip(CircleShape)
-                .background(MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.8f))
-                .clickable(
-                    interactionSource = remember { MutableInteractionSource() },
-                    indication = null
-                ) {
-                    isFullscreen = true
-                },
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                imageVector = Icons.Default.Fullscreen,
-                contentDescription = stringResource(R.string.fullscreen),
-                modifier = Modifier.size(18.dp),
-                tint = MaterialTheme.colorScheme.onSurfaceVariant
+    FullscreenChartFrame(
+        modifier = modifier,
+        hasTimeLabels = timeLabels.isNotEmpty(),
+        inline = {
+            OptimizedLineChart(
+                data = data,
+                color = color,
+                unit = unit,
+                showZeroLine = showZeroLine,
+                fixedMinMax = fixedMinMax,
+                timeLabels = timeLabels,
+                convertValue = convertValue,
+                externalSelectedFraction = externalSelectedFraction,
+                onXSelected = onXSelected,
+                fractionToTimeLabel = fractionToTimeLabel,
+                annotationRanges = annotationRanges,
+                modifier = Modifier.fillMaxWidth()
             )
-        }
-    }
-
-    // Fullscreen overlay
-    if (isFullscreen) {
-        FullscreenChartOverlay(
-            data = data,
-            color = color,
-            unit = unit,
-            showZeroLine = showZeroLine,
-            fixedMinMax = fixedMinMax,
-            timeLabels = timeLabels,
-            convertValue = convertValue,
-            annotationRanges = annotationRanges,
-            activity = activity,
-            onDismiss = { isFullscreen = false }
-        )
-    }
-}
-
-@Composable
-private fun FullscreenChartOverlay(
-    data: List<Float>,
-    color: Color,
-    unit: String,
-    showZeroLine: Boolean,
-    fixedMinMax: Pair<Float, Float>?,
-    timeLabels: List<String>,
-    convertValue: (Float) -> Float,
-    annotationRanges: List<AnnotationRange>,
-    activity: Activity?,
-    onDismiss: () -> Unit
-) {
-    val view = LocalView.current
-
-    // Lock orientation to landscape and hide system bars
-    DisposableEffect(Unit) {
-        activity?.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_USER_LANDSCAPE
-
-        // Hide system bars on the activity window
-        activity?.window?.let { window ->
-            WindowCompat.setDecorFitsSystemWindows(window, false)
-            val controller = WindowInsetsControllerCompat(window, view)
-            controller.hide(WindowInsetsCompat.Type.systemBars())
-            controller.systemBarsBehavior =
-                WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
-        }
-
-        onDispose {
-            // Restore portrait mode and show system bars
-            activity?.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
-            activity?.window?.let { window ->
-                WindowCompat.setDecorFitsSystemWindows(window, true)
-                val controller = WindowInsetsControllerCompat(window, view)
-                controller.show(WindowInsetsCompat.Type.systemBars())
-            }
-        }
-    }
-
-    // Handle back button press
-    BackHandler {
-        onDismiss()
-    }
-
-    Dialog(
-        onDismissRequest = onDismiss,
-        properties = DialogProperties(
-            dismissOnBackPress = true,
-            dismissOnClickOutside = false,
-            usePlatformDefaultWidth = false,
-            decorFitsSystemWindows = false,
-            securePolicy = SecureFlagPolicy.Inherit
-        )
-    ) {
-        BoxWithConstraints(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.surface)
-        ) {
-            // Calculate chart height based on available space minus padding and time labels
-            val verticalPadding = 48.dp // top + bottom padding
-            val timeLabelHeight = if (timeLabels.isNotEmpty()) 20.dp else 0.dp
-            val availableChartHeight = maxHeight - verticalPadding - timeLabelHeight
-
-            // Fullscreen chart with padding
+        },
+        fullscreen = { chartHeight ->
             OptimizedLineChart(
                 data = data,
                 color = color,
@@ -202,33 +58,9 @@ private fun FullscreenChartOverlay(
                 timeLabels = timeLabels,
                 convertValue = convertValue,
                 annotationRanges = annotationRanges,
-                chartHeight = availableChartHeight,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        start = 56.dp,  // Space for back button
-                        end = 24.dp,
-                        top = 24.dp,
-                        bottom = 24.dp
-                    )
+                chartHeight = chartHeight,
+                modifier = Modifier.fillMaxWidth()
             )
-
-            // Back button in top-left corner
-            IconButton(
-                onClick = onDismiss,
-                modifier = Modifier
-                    .align(Alignment.TopStart)
-                    .padding(8.dp)
-                    .size(40.dp)
-                    .clip(CircleShape)
-                    .background(MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.9f))
-            ) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = stringResource(R.string.exit_fullscreen),
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
         }
-    }
+    )
 }
